Fix missing lastUpdate season, that was a pain to track down. Not sure why I removed it.

+++ protocol/contracts/beanstalk/silo/WhitelistFacet.sol
@@ -17,7 +17,7 @@ contract WhitelistFacet {
     event WhitelistToken(
         address indexed token,
         bytes4 selector,
-        uint256 stalkPerBdvPerSeason,
+        uint32 stalkPerBdvPerSeason,
         uint256 stalk
     );
